Migrated 'align_images_using_orb' and its associated tests into their own modules
